add order queue functionality
