+++ frontend/src/App.jsx
@@ -5,34 +5,45 @@ import Register from "./pages/Register"
 import Home from "./pages/Home"
 import NotFound from "./pages/NotFound"
 import ProtectedRoute from "./components/ProtectedRoute"
+import ActiveOrders from "./pages/ActiveOrders"
+import Menu from "./pages/Menu"
+import OrderHistory from "./pages/OrderHistory"
 
-function Logout(){
+function Logout() {
   localStorage.clear()
   return <Navigate to="/login" />
 }
 
-function RegisterAndLogout(){
+function RegisterAndLogout() {
   localStorage.clear()
   return <Register />
 }
 
 function App() {
   return (
-   <BrowserRouter>
-    <Routes>
-      <Route  
-        path="/" 
-        element={
-        <ProtectedRoute>
-          <Home />
-        </ProtectedRoute>} 
-      />
-      <Route path="/login" element={<Login />}/>
-      <Route path="/logout" element={<Logout />}/>
-      <Route path="/register" element={<RegisterAndLogout />}/>
-      <Route path="*" element={<NotFound />}></Route>
-    </Routes> 
-   </BrowserRouter>
+    <BrowserRouter>
+      <Routes>
+        <Route
+          path="/"
+          element={
+            <ProtectedRoute>
+              <Home />
+            </ProtectedRoute>
+          }
+        >
+          {/* Nested routes rendered in <Outlet /> inside Home */}
+          <Route index element={<ActiveOrders />} />
+          <Route path="active-orders" element={<ActiveOrders />} />
+          <Route path="menu" element={<Menu />} />
+          <Route path="order-history" element={<OrderHistory />} />
+        </Route>
+
+        <Route path="/login" element={<Login />} />
+        <Route path="/logout" element={<Logout />} />
+        <Route path="/register" element={<RegisterAndLogout />} />
+        <Route path="*" element={<NotFound />} />
+      </Routes>
+    </BrowserRouter>
   )
 }
