+++ frontend/src/components/Note.jsx
@@ -18,7 +18,7 @@ function Note({note, onDelete}){
         <p className="note-content">{note.content}</p>
         <p className="note-date">{formattedDate}</p>
         <button className="delete-button" onClick={() => onDelete(note.id)}>
-            Delete
+            Done
         </button>
     </div>
 }
