+++ frontend/src/pages/Menu.jsx
@@ -1,0 +1,9 @@
+const Menu = () => {
+    return(
+        <>
+            <div>hello</div>
+        </>
+    )
+}
+
+export default Menu
