+++ frontend/src/pages/OrderHistory.jsx
@@ -1,0 +1,9 @@
+const OrderHistory = () => {
+    return(
+        <>
+            <div>hello</div>
+        </>
+    )
+}
+
+export default OrderHistory
