+++ frontend/src/style/Sidebar.css
@@ -1,0 +1,48 @@
+.sidebar {
+    width: 220px;
+    height: 100vh;
+    background-color: #3b2f2f; /* deep mocha brown */
+    padding: 2rem 1.5rem;
+    color: #fff3e0; /* creamy white */
+    position: fixed;
+    top: 0;
+    left: 0;
+    font-family: 'Georgia', serif;
+    box-shadow: 2px 0 5px rgba(0, 0, 0, 0.1);
+  }
+  
+  .sidebar .logo {
+    font-size: 1.8rem;
+    margin-bottom: 2.5rem;
+    color: #ffcb74; /* warm caramel */
+    font-family: 'Pacifico', cursive; /* Try importing this font if you can */
+    text-align: center;
+  }
+  
+  .sidebar nav {
+    display: flex;
+    flex-direction: column;
+    gap: 1.5rem;
+  }
+  
+  .sidebar a {
+    text-decoration: none;
+    color: #e0c8b0; /* light coffee cream */
+    font-weight: 500;
+    font-size: 1.1rem;
+    padding: 0.5rem 1rem;
+    border-radius: 12px;
+    transition: all 0.2s ease;
+  }
+  
+  .sidebar a:hover {
+    background-color: #5c4438;
+    color: #fff3e0;
+  }
+  
+  .sidebar a.active {
+    background-color: #ffcb74;
+    color: #3b2f2f;
+    font-weight: bold;
+  }
+
